feat: make sure resource providers are initialized in k8s context

+++ opal-r/src/main/java/org/obiba/opal/r/kubernetes/RockPodSession.java
@@ -18,6 +18,8 @@ public class RockPodSession extends RockSession {
 
   private final PodRef pod;
 
+  private boolean terminatePod = true;
+
   protected RockPodSession(String serverName, PodRef pod, AppCredentials credentials, String user, PodsService podsService, TransactionalThreadFactory transactionalThreadFactory, EventBus eventBus) throws RServerException {
     super(serverName, pod.getName(), credentials, user, transactionalThreadFactory, eventBus);
     this.pod = pod;
@@ -25,12 +27,16 @@ public class RockPodSession extends RockSession {
     openSession();
   }
 
+  public void setTerminatePod(boolean terminatePod) {
+    this.terminatePod = terminatePod;
+  }
+
   @Override
   public void close() {
     if (isClosed()) return;
     super.close();
     // terminate pod
-    podsService.deletePod(pod);
+    if (terminatePod) podsService.deletePod(pod);
   }
 
   public PodRef getPod() {
